refactor(commands): split oversized command files
Both command files sat at the 300-350 soft ceiling. Extract
cohesive groups into their own files, all now within the 100-250
sweet spot. Pure move — no behavior change.

- KiwiCore+Commands.swift (312 -> 230): navigate -> new
  KiwiCore+NavigateCommand.swift.
- KiwiCore+LayoutCommands.swift (350 -> 199): scroll.* -> new
  KiwiCore+ScrollCommands.swift; global toggles -> new
  KiwiCore+SettingsCommands.swift.
- parsePlacement / placementError widened private -> internal so
  the split-out files reuse them.
